Move EdgeAnnoSearchSpec from aql to more general operator package
