+++ src/aql/operators/edge_op.rs
@@ -2,103 +2,12 @@ use annostorage::AnnoStorage;
 use {AnnoKey, Annotation, Component, ComponentType, Edge, Match, NodeID};
 use graphstorage::{GraphStatistic, GraphStorage};
 use graphdb::{GraphDB, ANNIS_NS};
-use operator::{EstimationType, Operator, OperatorSpec};
+use operator::{EstimationType, Operator, OperatorSpec, EdgeAnnoSearchSpec};
 use util;
 use std;
 use std::collections::VecDeque;
 use std::sync::Arc;
 use stringstorage::StringStorage;
-
-#[derive(Clone, Debug)]
-pub enum EdgeAnnoSearchSpec {
-    ExactValue {
-        ns: Option<String>,
-        name: String,
-        val: Option<String>,
-    },
-}
-
-impl std::fmt::Display for EdgeAnnoSearchSpec {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        match self {
-            &EdgeAnnoSearchSpec::ExactValue {
-                ref ns,
-                ref name,
-                ref val,
-            } => {
-                let qname = if let &Some(ref ns) = ns {
-                    format!("{}:{}", ns, name)
-                } else {
-                    name.clone()
-                };
-
-                if let &Some(ref val) = val {
-                    write!(f, "{}={}", qname, val)
-                } else {
-                    write!(f, "{}", qname)
-                }
-            }
-        }
-    }
-}
-
-impl EdgeAnnoSearchSpec {
-    pub fn get_anno(&self, strings: &StringStorage) -> Option<Annotation> {
-        match self {
-            &EdgeAnnoSearchSpec::ExactValue {
-                ref ns,
-                ref name,
-                ref val,
-            } => {
-                let ns = if let &Some(ref s) = ns {
-                    strings.find_id(s)?.clone()
-                } else {
-                    0
-                };
-                let val = if let &Some(ref s) = val {
-                    strings.find_id(s)?.clone()
-                } else {
-                    0
-                };
-                let name = strings.find_id(name)?.clone();
-                let mut anno = Annotation {
-                    key: AnnoKey { ns, name },
-                    val,
-                };
-
-                Some(anno)
-            }
-        }
-    }
-
-    pub fn guess_max_count(
-        &self,
-        anno_storage: &AnnoStorage<Edge>,
-        strings: &StringStorage,
-    ) -> Option<usize> {
-        match self {
-            &EdgeAnnoSearchSpec::ExactValue {
-                ref ns,
-                ref name,
-                ref val,
-            } => {
-                let val = val.clone()?;
-                let name_id = strings.find_id(&name)?;
-                if let Some(ns) = ns.clone() {
-                    let ns_id = strings.find_id(&ns)?;
-                    return Some(anno_storage.guess_max_count(
-                        Some(ns_id.clone()),
-                        name_id.clone(),
-                        &val,
-                        &val,
-                    ));
-                } else {
-                    return Some(anno_storage.guess_max_count(None, name_id.clone(), &val, &val));
-                }
-            }
-        }
-    }
-}
 
 #[derive(Clone, Debug)]
 struct BaseEdgeOpSpec {
